Move WanPublisherState to wan package from config
Moving `WanPublisherState` from `com.hazelcast.config` to
`com.hazelcast.wan`. The enum is used to configure initial publisher
state of `WanBatchReplicationPublisherConfig`. Besides that, the same
enum is used internally the WAN implementation and on
`LocalWanPublisherStats` too, therefore it fits better in the wan
package than in config.
