+++ hazelcast/src/main/java/com/hazelcast/internal/management/dto/WanBatchReplicationPublisherConfigDTO.java
@@ -26,7 +26,7 @@ import com.hazelcast.config.KubernetesConfig;
 import com.hazelcast.config.WanQueueFullBehavior;
 import com.hazelcast.config.WanAcknowledgeType;
 import com.hazelcast.config.WanBatchReplicationPublisherConfig;
-import com.hazelcast.config.WanPublisherState;
+import com.hazelcast.wan.WanPublisherState;
 import com.hazelcast.config.WanSyncConfig;
 import com.hazelcast.internal.json.JsonObject;
 import com.hazelcast.internal.json.JsonValue;
