+++ hazelcast/src/main/java/com/hazelcast/internal/management/operation/ChangeWanStateOperation.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.internal.management.operation;
 
-import com.hazelcast.config.WanPublisherState;
+import com.hazelcast.wan.WanPublisherState;
 import com.hazelcast.spi.impl.operationservice.AbstractLocalOperation;
 import com.hazelcast.spi.impl.NodeEngine;
 import com.hazelcast.wan.impl.WanReplicationService;
