+++ hazelcast/src/test/java/com/hazelcast/internal/management/ChangeWanStateRequestTest.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.internal.management;
 
-import com.hazelcast.config.WanPublisherState;
+import com.hazelcast.wan.WanPublisherState;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.internal.management.request.ChangeWanStateRequest;
 import com.hazelcast.internal.json.JsonObject;
@@ -29,9 +29,9 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import static com.hazelcast.config.WanPublisherState.PAUSED;
-import static com.hazelcast.config.WanPublisherState.REPLICATING;
-import static com.hazelcast.config.WanPublisherState.STOPPED;
+import static com.hazelcast.wan.WanPublisherState.PAUSED;
+import static com.hazelcast.wan.WanPublisherState.REPLICATING;
+import static com.hazelcast.wan.WanPublisherState.STOPPED;
 import static com.hazelcast.internal.util.JsonUtil.getString;
 import static org.junit.Assert.assertNotEquals;
